Add support for loading config from a clcache.conf file
This is useful and the GCC version of ccache also implements a similar concept.

It might be useful in situations where some default settings could be
specified or when environment is not passed correctly down to the compiler
chain.

This change introduces support for clcache.conf file to allow clcache
to read settings from there.

The precedence is as follows:
 1. If present, Environment variable(s) are used, or
 2. [current_working_dir]\clcache.conf is loaded, or, if not found
 3. [%HOME% or ~\].clcache\clcache.conf is be loaded, or if not found
 4. [%ALLUSERSPROFILE% or C:\Users\].clcache\clcache.conf is loaded.

In each case, once a clcache.conf file is found, no other conf file is
considered and values only from this file are used. It is also loaded
only once (and all its values then cached) - all to avoid unnecessary
performance penalties.